fixed scrolling glitch
