+++ src/App.css
@@ -4,5 +4,5 @@
 }
 
 html {
-  scroll-snap-type: y proximity;
+  scroll-snap-type: y mandatory;
 }
